+++ src/Smartstore.Web.Common/WebStarter.cs
@@ -155,6 +155,7 @@ namespace Smartstore.Web
             builder.Configure(StarterOrdering.AfterStaticFilesMiddleware, app =>
             {
                 app.UsePoweredBy();
+                app.UseSecurityHeaders();
             });
 
             builder.Configure(StarterOrdering.RoutingMiddleware, app =>